+++ src/lib.rs
@@ -2,6 +2,7 @@ pub mod display;
 pub mod file;
 pub mod hello;
 pub mod images;
+pub mod job;
 pub mod path;
 pub mod rwlock;
 pub mod sqlite;